# return reason for allowed rbac authorizations