@enj 2018-01-19
@liggitt why not use more structured output instead of a giant string?