@liggitt 2018-01-19
> @liggitt why not use more structured output instead of a giant string?

* we already have plumbing for a string reason
* putting structure inside a string is fragile and makes the interface misleading
* the same interface needs to be remoteable (via subject access review)
* the intended consumer of this (audit) is unlikely to nest structured data from arbitrary authorizers. see https://github.com/kubernetes/kubernetes/pull/58143 for discussion of a way for authorizers/admission plugins to contribute unstructured attributes to an audit event